Added examples in functions updateUser, updateProduct, login and checkRegister
function login: Checks if the provided username exists and verifies the corresponding password. If successful, it returns true; otherwise, false.

function checkRegister: Checks if a given username exists in the database table "tblUser" by executing a prepared SQL statement. If the username is not found, it returns true; otherwise, false.

function updateProduct: Updates the user information in the "tblUser" table with the provided data, based on the user's ID.

function updateProduct: Updates the product information in the "tblProduct" table with the provided data, based on the product's ID.